Tests fix; move synonym parsers as well
