+++ modules/analysis-common/src/main/java/org/elasticsearch/analysis/common/ESSolrSynonymParser.java
@@ -17,23 +17,23 @@
  * under the License.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
 import org.apache.lucene.analysis.Analyzer;
-import org.apache.lucene.analysis.synonym.WordnetSynonymParser;
+import org.apache.lucene.analysis.synonym.SolrSynonymParser;
 import org.apache.lucene.util.CharsRef;
 import org.apache.lucene.util.CharsRefBuilder;
 
 import java.io.IOException;
 
-public class ESWordnetSynonymParser extends WordnetSynonymParser {
-    private static final Logger logger = LogManager.getLogger(ESWordnetSynonymParser.class);
+public class ESSolrSynonymParser extends SolrSynonymParser {
+    private static final Logger logger = LogManager.getLogger(ESSolrSynonymParser.class);
 
     private final boolean lenient;
 
-    public ESWordnetSynonymParser(boolean dedup, boolean expand, boolean lenient, Analyzer analyzer) {
+    public ESSolrSynonymParser(boolean dedup, boolean expand, boolean lenient, Analyzer analyzer) {
         super(dedup, expand, analyzer);
         this.lenient = lenient;
     }
