+++ modules/analysis-common/src/test/java/org/elasticsearch/analysis/common/ESSolrSynonymParserTests.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.elasticsearch.index.analysis;
+package org.elasticsearch.analysis.common;
 
 import org.apache.lucene.analysis.CharArraySet;
 import org.apache.lucene.analysis.StopFilter;
@@ -35,16 +35,13 @@ import java.text.ParseException;
 
 import static org.hamcrest.Matchers.containsString;
 
-public class ESWordnetSynonymParserTests extends ESTokenStreamTestCase {
+public class ESSolrSynonymParserTests extends ESTokenStreamTestCase {
 
     public void testLenientParser() throws IOException, ParseException {
-        ESWordnetSynonymParser parser = new ESWordnetSynonymParser(true, false, true, new StandardAnalyzer());
+        ESSolrSynonymParser parser = new ESSolrSynonymParser(true, false, true, new StandardAnalyzer());
         String rules =
-            "s(100000001,1,'&',a,1,0).\n" +
-            "s(100000001,2,'and',a,1,0).\n" +
-            "s(100000002,1,'come',v,1,0).\n" +
-            "s(100000002,2,'advance',v,1,0).\n" +
-            "s(100000002,3,'approach',v,1,0).";
+            "&,and\n" +
+            "come,advance,approach\n";
         StringReader rulesReader = new StringReader(rules);
         parser.parse(rulesReader);
         SynonymMap synonymMap = parser.build();
@@ -57,12 +54,9 @@ public class ESWordnetSynonymParserTests extends ESTokenStreamTestCase {
     public void testLenientParserWithSomeIncorrectLines() throws IOException, ParseException {
         CharArraySet stopSet = new CharArraySet(1, true);
         stopSet.add("bar");
-        ESWordnetSynonymParser parser =
-            new ESWordnetSynonymParser(true, false, true, new StandardAnalyzer(stopSet));
-        String rules =
-            "s(100000001,1,'foo',v,1,0).\n" +
-            "s(100000001,2,'bar',v,1,0).\n" +
-            "s(100000001,3,'baz',v,1,0).";
+        ESSolrSynonymParser parser =
+            new ESSolrSynonymParser(true, false, true, new StandardAnalyzer(stopSet));
+        String rules = "foo,bar,baz";
         StringReader rulesReader = new StringReader(rules);
         parser.parse(rulesReader);
         SynonymMap synonymMap = parser.build();
@@ -73,16 +67,12 @@ public class ESWordnetSynonymParserTests extends ESTokenStreamTestCase {
     }
 
     public void testNonLenientParser() {
-        ESWordnetSynonymParser parser = new ESWordnetSynonymParser(true, false, false, new StandardAnalyzer());
+        ESSolrSynonymParser parser = new ESSolrSynonymParser(true, false, false, new StandardAnalyzer());
         String rules =
-            "s(100000001,1,'&',a,1,0).\n" +
-            "s(100000001,2,'and',a,1,0).\n" +
-            "s(100000002,1,'come',v,1,0).\n" +
-            "s(100000002,2,'advance',v,1,0).\n" +
-            "s(100000002,3,'approach',v,1,0).";
+            "&,and=>and\n" +
+            "come,advance,approach\n";
         StringReader rulesReader = new StringReader(rules);
         ParseException ex = expectThrows(ParseException.class, () -> parser.parse(rulesReader));
         assertThat(ex.getMessage(), containsString("Invalid synonym rule at line 1"));
     }
-
 }
